feature: SignIn and SignOut buttons
