+++ configs/auth.ts
