+++ src/app/components/Header/SignOutButton.tsx
@@ -1,0 +1,15 @@
+import { signOut } from 'next-auth/react'
+import React from 'react'
+
+const SignOutButton = () => {
+  return (
+    <button
+        className='userButton'
+        onClick={() => signOut({callbackUrl: '/'})}
+    >
+        Sign Out
+    </button>
+  )
+}
+
+export default SignOutButton
