+++ src/app/modules/Header/application/HeaderContainer/page.tsx
@@ -18,7 +18,9 @@ export default function HeaderContainer({ session }: ISession) {
 
                 </ul>
             </div>
+
             <HeaderAuthSection session={session}/>
+            
         </nav>
     )
 }
